Sorted out sizes, and masses for creatures/body parts.  Added a ton of integration tests, using data pulled from df hack, that we are hitting to acceptable accuracy

+++ Tiles/DriverConsole/Tiles.Content.Bridge.DfNet/IDfAgentBuilder.cs
@@ -28,7 +28,7 @@ namespace Tiles.Content.Bridge.DfNet
 
         void OverrideBodyPartCategorySize(string bpCategory, int size);
 
-        void AddLifeStageSize(int ageYear, int ageDay, int size);
+        void AddLifeStageSize(int ageYear, int ageDay, double size);
 
         void AddBodyAttack(DfBodyAttack move);
     }
